refactor(self): move github manipulations to the appropriate github.py module
